Camera3: Increase mOfflineMetaMemory count to account for reprocess pipelining
The mOfflineMetaMemory is used to map the metadata for an input reprocess
request which includes both pproc + jpeg. However, we unblock the reprocess
request after the pproc, which means that we could get another reprocess
request while jpeg is ongoing and the mOfflineMetaMemory is used. So add
another buffer to the metaMemory to allow the pipelining.

MAX_REPROCESS_PIPELINE_STAGES = pproc + jpeg = 2
Bugs-fixed: 24141064

Ack by tgodbole

Bugs: 24141064
Change-Id: I44c3464c11431f4a654c699471aaab33a818ba78